requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==2.0.20
alembic==1.18.0
pydantic==1.10.12
python-multipart==0.0.6
requests==2.31.0
httpx==0.24.1
python-dotenv==1.0.1
google-auth==2.24.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.95.0
python-docx==0.8.11
pdfplumber==0.8.0
openai==1.0.0
aiofiles==23.1.0
cryptography==41.0.3
itsdangerous==2.1.2
bcrypt==4.0.1

